Changing the map-generating/transmission structure, doing some general cleaning/refactoring. Next to do: fill out Client and Server OnMapCheck handlers, add a "map waiting list", and retool the Server MapTransmit handler.

+++ Assets/Scripts/Unit.cs
@@ -14,7 +14,7 @@ public class Unit : MonoBehaviour, ITurnable, ISelectable
     public Delegate Ability04;
     public Delegate Ability05;
     public ShipScale Scale;
-    public float MaxHealth, CurrentHealth;
+    public int MaxHealth, CurrentHealth;
     public List<HexTile> TilesInRange
     {
         get
